Moves prosecutor change handling to component
