+++ apps/judicial-system/web/src/routes/Prosecutor/SharedComponents/ProsecutorSection/ProsecutorSection.tsx
@@ -1,20 +1,34 @@
-import React from 'react'
+import React, { useContext } from 'react'
 
 import { User } from '@island.is/judicial-system/types'
+import { FormContext } from '@island.is/judicial-system-web/src/components/FormProvider/FormProvider'
+import { useCase } from '@island.is/judicial-system-web/src/utils/hooks'
 import ProsecutorSectionHeading from './ProsecutorSectionHeading'
 import ProsecutorSelection from './ProsecutorSelection'
 
-interface Props {
-  onChange: (prosecutor: User) => boolean
-}
+const ProsecutorSection: React.FC = () => {
+  const { workingCase, setWorkingCase } = useContext(FormContext)
+  const { setAndSendToServer } = useCase()
 
-const ProsecutorSection: React.FC<Props> = (props) => {
-  const { onChange } = props
+  const handleProsecutorChange = (prosecutor: User) => {
+    setAndSendToServer(
+      [
+        {
+          prosecutorId: prosecutor.id,
+          force: true,
+        },
+      ],
+      workingCase,
+      setWorkingCase,
+    )
+
+    return true
+  }
 
   return (
     <>
       <ProsecutorSectionHeading />
-      <ProsecutorSelection onChange={onChange} />
+      <ProsecutorSelection onChange={handleProsecutorChange} />
     </>
   )
 }
